third_party/WebKit/LayoutTests/crypto/subtle/hmac/legacy-empty-key-expected.txt

Tests structured de-cloning of empty HMAC keys

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Deserializing empty HMAC SHA-1 key...
PASS key.type is "secret"
PASS typeof key.extractable is 'boolean'
PASS key.extractable is true
PASS key.algorithm.name is "HMAC"
PASS key.algorithm.length is 0
PASS key.algorithm.hash.name is "SHA-1"
PASS key.usages is ["sign", "verify"]

Deserializing empty HMAC SHA-256 key...
PASS key.type is "secret"
PASS typeof key.extractable is 'boolean'
PASS key.extractable is true
PASS key.algorithm.name is "HMAC"
PASS key.algorithm.length is 0
PASS key.algorithm.hash.name is "SHA-256"
PASS key.usages is ["sign", "verify"]

calling verify() with a valid signature (SHA-1) ...
PASS typeof verifyResult is 'boolean'
PASS verifyResult is true

calling verify() with an invalid signature (SHA-1) ...
PASS typeof verifyResult is 'boolean'
PASS verifyResult is false

calling sign() (SHA-1) over empty input...
PASS: signature should be [fbdb1d1b18aa6c08324b7d64b71fb76370690e1d] and was

calling verify() with a valid signature (SHA-256) ...
PASS typeof verifyResult is 'boolean'
PASS verifyResult is true

calling verify() with an invalid signature (SHA-256) ...
PASS typeof verifyResult is 'boolean'
PASS verifyResult is false

calling sign() (SHA-256) over empty input...
PASS: signature should be [b613679a0814d9ec772f95d778c35fc5ff1697c493715653c6c712144292c5ad] and was
PASS successfullyParsed is true

TEST COMPLETE
